+++ lib/ReactViews/Story/StoryPanel.jsx
@@ -170,7 +170,7 @@ const StoryPanel = observer(
               `Story data: \`${story.title ?? story.id}\``,
               // We set errors to use Warning severity so they aren't shown to the user by default
               // This is due to many stories/shareData having invalid models in them
-              // If a more severe error is thrown while loading shareData (eg Error or Critical) then the error WILL still be shown to the user
+              // If a more severe error is thrown while loading shareData (eg Error) then the error WILL still be shown to the user
               TerriaErrorSeverity.Warning
             )
             .catch(function(e) {